Add ability to generate png with inframap
The APIs for map generation via inframap are updated to support
generating png representations of the map using GraphViz's `dot` command.
For this purpose, graphviz is installed in the Docker image, and an
option struct is added to `api.InfraMap()`.

The command line interface is slightly refactored to allow for
tool-specific options in order to support this new feature.

The ability to generate png is available both via the command line and
the REST API. The /map endpoint will accept a "png" boolean attribute in
the JSON body.