Workaround inframap's icon generation issue
Inframap sometimes generates empty graphs (as detailed here:
https://github.com/cycloidio/inframap#why-is-my-graph-generated-empty).

Validiac prefers to generate graphs with icons so that they are more
pleasing to the eye. The previous commit added the `--raw` flag for all
inframap usage, which prevents icons from being generated, but also
basically ensures a non-empty graph is returned.

Since HCL/state files used with validiac come from third parties, it is
difficult for us to understand why inframap is returning these empty
graphs.

To workaround this issue in the meantime, when generating pngs, validiac
will first inspect whether an empty graph has been generated, and if so,
rerun with --raw. This ensures that graphs that can be generated with
icons will be, and only graphs that can't are generated in raw mode.