Return only stdout of dot
When validiac executed dot (and other tools), it would take both their
stdout and stderr outputs. For other tools, this is fine because they
print to stderr when input files are invalid (i.e. fail their checks),
but for dot this could be a problem: when the graph is very big, dot
prints a warning to stderr. This basically corrupted the generated png.

Now, only stdout of dot will be taken. Its stderr will be returned only
if dot failed.